T1: ArrayPeriodIterator implemented, partially tested; T2: Filter with predicate parametrizated, Group, Student, implemented, partially tested

+++ BT-Library-Java/src/bt/core/collections/elems/Player.java
@@ -2,14 +2,14 @@ package bt.core.collections.elems;
 
 public class Player {
 	private final String name;
-	private final int points;
+	private final double points;
 	
-	public Player (String name, int points){
+	public Player (String name, double points){
 		this.name = name;
 		this.points = points;
 	}
 	
-	public int getPoints() {
+	public double getPoints() {
 		return points;
 	}
